# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BioGene Unified Bioinformatics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nlp/extract:
    post:
      operationId: extractBiomedicalEntities
      tags: [nlp]
      summary: Extract biomedical entities from text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NlpExtractBody"
      responses:
        "200":
          description: Extracted entities and relations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NlpExtractResult"
  /nlp/papers/search:
    get:
      operationId: searchPubmedPapers
      tags: [nlp]
      summary: Search PubMed for open-access papers
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Papers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PubmedSearchResponse"
  /nlp/knowledge-graph:
    get:
      operationId: getKnowledgeGraph
      tags: [nlp]
      summary: Get the current knowledge graph of entities and relationships
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Knowledge graph data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraph"

  # ── Genome Browser Module ──────────────────────────────────────────────────
...
components:
  schemas:
...
    BioEntity:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        type:
          type: string
        normalizedId:
          type: string
        confidence:
          type: number
        startOffset:
          type: integer
        endOffset:
          type: integer
      required: [id, text, type, confidence]
    EntityRelation:
      type: object
      properties:
        subject:
          type: string
        predicate:
          type: string
        object:
          type: string
        confidence:
          type: number
        evidence:
          type: string
      required: [subject, predicate, object, confidence]
    NlpExtractBody:
      type: object
      properties:
        text:
          type: string
        pmid:
          type: string
      required: []
    NlpExtractResult:
      type: object
      properties:
        entities:
          type: array
          items:
            $ref: "#/components/schemas/BioEntity"
        relations:
          type: array
          items:
            $ref: "#/components/schemas/EntityRelation"
        text:
          type: string
        entityCounts:
          type: object
          additionalProperties:
            type: integer
        source:
          type: string
        pmid:
          type: string
        pubmedUrl:
          type: string
        note:
          type: string
      required: [entities, relations, text]
    PubmedPaper:
      type: object
      properties:
        pmid:
          type: string
        title:
          type: string
        abstract:
          type: string
        authors:
          type: array
          items:
            type: string
        journal:
          type: string
        year:
          type: integer
        doi:
          type: string
        keywords:
          type: array
          items:
            type: string
        citationCount:
          type: integer
        pubmedUrl:
          type: string
        doiUrl:
          type: string
      required: [pmid, title]
    PubmedSearchResponse:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/PubmedPaper"
        total:
          type: integer
        query:
          type: string
      required: [papers, total, query]
    KnowledgeGraphNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        count:
          type: integer
      required: [id, label, type]
    KnowledgeGraphEdge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        relation:
          type: string
        confidence:
          type: number
      required: [source, target, relation, confidence]
    KnowledgeGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeGraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeGraphEdge"
        totalDocuments:
          type: integer
        entityTypeCounts:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
            required: [type, count]
      required: [nodes, edges, totalDocuments]

    # ── Genome Browser Schemas ─────────────────────────────────────────────────
...